Convert menu to a separate component file.
